Tests (#8)
## Proposed changes

Account Tests

## Types of changes

What types of changes does your code introduce?
_Put an `x` in the boxes that apply_

- [ ] Bugfix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Refactoring (improve a current implementation without adding a new feature or fixing a bug)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] Build (changes that affect the build system)
- [ ] Docs (documentation only changes)
- [x] Test (adding missing tests or fixing existing tests)
- [ ] Other... Please describe:

## Checklist

_Put an `x` in the boxes that apply. You can also fill these out after creating the PR. If you're unsure about any of them, don't hesitate to ask. We're here to help! This is simply a reminder of what we are going to look for before merging your code._

- [x] I have read the [CONTRIBUTING](https://docs.ark.io/docs/contributing) documentation
- [x] Lint and unit tests pass locally with my changes
- [x] I have added tests that prove my fix is effective or that my feature works
- [ ] I have added necessary documentation (if appropriate)

+++ client/one/blocks_requests.go
@@ -1,20 +1,22 @@
-// Copyright 2018 ArkEcosystem. All rights reserved.
+// This file is part of Ark Go Client.
 //
-// Use of this source code is governed by the MIT
-// license that can be found in the LICENSE file.
+// (c) Ark Ecosystem <info@ark.io>
+//
+// For the full copyright and license information, please view the LICENSE
+// file that was distributed with this source code.
 
 package one
 
 type BlocksQuery struct {
-	Limit              int    `url:"limit"`
+	Limit              byte   `url:"limit"`
 	OrderBy            string `url:"orderBy"`
-	Offset             int    `url:"offset"`
+	Offset             byte   `url:"offset"`
 	GeneratorPublicKey string `url:"generatorPublicKey"`
-	TotalAmount        int    `url:"totalAmount"`
-	TotalFee           int    `url:"totalFee"`
-	Reward             int    `url:"reward"`
+	TotalAmount        byte   `url:"totalAmount"`
+	TotalFee           byte   `url:"totalFee"`
+	Reward             byte   `url:"reward"`
 	PreviousBlock      string `url:"previousBlock"`
-	Height             int    `url:"height"`
+	Height             byte   `url:"height"`
 }
 
 type BlockIdQuery struct {
